feat(doc integrations): Add GET endpoint for all default IntegrationFeatures (#30769)
See API-2333

This PR adds an endpoint to view all the default IntegrationFeatures. Since it lives in Python and not the DB, My work around to "serialize" it was to create an IntegrationFeature in memory, and serialize that one without any FK fields. I could have just manually assembled the dictionary in the endpoint, or created a new serializer just for the Features enum, but I thought this was the best compromise.

+++ src/sentry/api/serializers/rest_framework/doc_integration.py
@@ -59,7 +59,7 @@ class DocIntegrationSerializer(Serializer):
 
     def create(self, validated_data: MutableMapping[str, Any]) -> DocIntegration:
         slug = self._generate_slug(validated_data["name"])
-        features = validated_data.pop("features")
+        features = validated_data.pop("features") if validated_data.get("features") else []
         with transaction.atomic():
             doc_integration = DocIntegration.objects.create(slug=slug, **validated_data)
             IntegrationFeature.objects.bulk_create(
@@ -77,20 +77,21 @@ class DocIntegrationSerializer(Serializer):
     def update(
         self, doc_integration: DocIntegration, validated_data: MutableMapping[str, Any]
     ) -> DocIntegration:
-        features = validated_data.pop("features")
-        # Delete any unused features
-        unused_features = IntegrationFeature.objects.filter(
-            target_id=doc_integration.id, target_type=IntegrationTypes.DOC_INTEGRATION.value
-        ).exclude(feature__in=features)
-        for unused_feature in unused_features:
-            unused_feature.delete()
-        # Create any new features
-        for feature in features:
-            IntegrationFeature.objects.get_or_create(
-                target_id=doc_integration.id,
-                target_type=IntegrationTypes.DOC_INTEGRATION.value,
-                feature=feature,
-            )
+        if validated_data.get("features"):
+            features = validated_data.pop("features")
+            # Delete any unused features
+            unused_features = IntegrationFeature.objects.filter(
+                target_id=doc_integration.id, target_type=IntegrationTypes.DOC_INTEGRATION.value
+            ).exclude(feature__in=features)
+            for unused_feature in unused_features:
+                unused_feature.delete()
+            # Create any new features
+            for feature in features:
+                IntegrationFeature.objects.get_or_create(
+                    target_id=doc_integration.id,
+                    target_type=IntegrationTypes.DOC_INTEGRATION.value,
+                    feature=feature,
+                )
         # Update the DocIntegration
         for key, value in validated_data.items():
             setattr(doc_integration, key, value)
